refactor: move parse of FDK specific feilds to seperate package and add common fields in base parser
