+++ src/main/kotlin/no/digdir/fdk/parseservice/extract/fdk/BuilderExtensions.kt
@@ -1,0 +1,20 @@
+package no.digdir.fdk.parseservice.extract.fdk
+
+import no.digdir.fdk.model.dataset.Dataset
+import no.digdir.fdk.parseservice.extract.containsTriple
+import no.digdir.fdk.parseservice.namespace.FDK
+import org.apache.jena.rdf.model.Resource
+
+/**
+ * Adds data to the dataset builder that has been added to the dataset graph by the FDK harvest process.
+ * @param recordResource The jena resource of the fdkRecord
+ * @param datasetResource The jena resource of the dataset
+ */
+fun Dataset.Builder.addFdkData(recordResource: Resource, datasetResource: Resource) {
+    setId(fdkIdFromRecord(recordResource))
+    setHarvest(harvestMetaData(recordResource))
+
+    setIsRelatedToTransportportal(datasetResource.model.containsTriple(datasetResource.uri, FDK.isRelatedToTransportportal.uri, true))
+    setIsOpenData(datasetResource.model.containsTriple(datasetResource.uri, FDK.isOpenData.uri, true))
+    setIsAuthoritative(datasetResource.model.containsTriple(datasetResource.uri, FDK.isAuthoritative.uri, true))
+}
